+++ pages/index.js
@@ -1,15 +1,8 @@
-import React, { useEffect, useRef } from 'react'
+import React from 'react'
 import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { defaults } from '../next-i18next.config'
 import Head from 'next/head'
-import WalletConnect from '@walletconnect/client'
-import QRCodeModal from 'algorand-walletconnect-qrcode-modal'
-
-const connector = new WalletConnect({
-  bridge: 'https://bridge.walletconnect.org', // Required
-  qrcodeModal: QRCodeModal,
-})
 
 // Material UI components
 import Container from '@mui/material/Container'
@@ -36,24 +29,8 @@ export async function getServerSideProps({ locale }) {
 }
 export default function Home() {
   const { t } = useTranslation('index')
-  const { addresses, activeWallet, setActiveWallet } = useRewardsAddresses()
+  const { addresses, activeWallet } = useRewardsAddresses()
   const isConnected = addresses.length > 0
-  const connectorRef = useRef(connector)
-
-  useEffect(() => {
-    // eslint-disable-next-line max-len
-    // This useEffect is necessary because when getting the wallet from localStorage the sendCustomRequest method is undefined
-    // rerunning peraConnect reAttaches the signing method to the connector.
-    if (
-      activeWallet?.type === 'wallet-connect' &&
-      typeof activeWallet.connector.sendCustomRequest === 'undefined'
-    ) {
-      setActiveWallet({
-        ...activeWallet,
-        connector: connectorRef.current,
-      })
-    }
-  }, [activeWallet])
 
   return (
     <>
